Correcciones
Control de usuarios, error en eliminar categorias, subcategorias y tipos

+++ proyecto/eliminar_tipo.php
@@ -17,6 +17,28 @@ if(isset($_POST['id'])) {
         die("Conexión fallida: " . $conn->connect_error);
     }
 
+     // Verificar si hay preguntas asociadas a esta categoría
+     $sql_check_questions = "SELECT COUNT(*) AS total FROM preguntas WHERE tipo_pregunta_id = $id";
+     $result_check_questions = $conn->query($sql_check_questions);
+ 
+     if ($result_check_questions === FALSE) {
+         // Enviar una respuesta al cliente indicando que ocurrió un error al verificar las preguntas asociadas
+         echo json_encode(array('success' => false, 'error' => 'Error al verificar las preguntas asociadas.'));
+ 
+         exit();
+     }
+ 
+     $row = $result_check_questions->fetch_assoc();
+     $total_questions = $row['total'];
+ 
+     if ($total_questions > 0) {
+         // Enviar una respuesta al cliente indicando que no se puede eliminar la categoría debido a preguntas asociadas
+         alert("No se puede eliminar el tipo pregunta porque tiene preguntas asociadas.");
+         echo json_encode(array('success' => false, 'error' => 'No se puede eliminar la categoría porque tiene preguntas asociadas.'));
+         
+         exit();
+     }
+
     // Preparar la consulta SQL para eliminar la categoría
     $sql = "DELETE FROM tipospreguntas WHERE id = $id";
